Finialize import function
[Jiri](https://fuchsia.googlesource.com/jiri) has `import` function, which imports other package's dependency into current worskpace.

Import function is only required once, due to the fact that workspace owner only needs to setup the initial dependency of the workspace.
When the dependency has new commits, the workspace only requires to update the top packages.

For instance, check the below graph.
```
top
\- open-source1
	\- open-source2
```
If open-source2 has new commits which should be applied into top but not for open-source1, the point of updating open-source1 dependency is regardless.
Moreover, if open-source1 is relatively slow on updating, the top package owner cannot applyt the new commits from open-sourc2 into his/her workspace.

Hence, use the below graph.
```
top
\- open-source1
	\- open-source2 (1)
\- open-source2 (2)
```
Initially the two open-source2 dependency points to the same commit. But once open-source2 is update, the user only requires to update open-source2 (2) dependency to the top package.